Updated naming to match install location
